requirements.txt
absl-py==0.10.0
aiohttp==3.6.3
aiohttp-socks==0.5.5
aiorpcX==0.18.4
asgiref==3.3.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
awsebcli==3.19.2
bcrypt==3.2.0
bitstring==3.1.7
blessed==1.17.11
botocore==1.19.20
cached-property==1.5.2
cachetools==4.1.1
cement==2.8.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3.1
cryptography==3.1.1
Django==3.1.4
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
dnspython==1.16.0
docker==4.3.1
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.20.0
ecdsa==0.16.0
Electrum==4.0.4
future==0.16.0
gast==0.3.3
google-auth==1.21.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
gunicorn==20.0.4
h5py==2.10.0
helpdev==0.7.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
Keras-Preprocessing==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
multidict==4.7.6
numpy==1.18.5
opt-einsum==3.3.0
packaging==20.8
paramiko==2.7.2
pathspec==0.5.9
protobuf==3.13.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-socks==1.1.0
pytz==2020.4
PyYAML==5.3.1
QDarkStyle==2.8.1
qrcode==6.1
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scipy==1.4.1
semantic-version==2.5.0
six==1.15.0
sqlparse==0.4.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
texttable==1.6.3
uritemplate==3.0.1
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.5.1
